+++ docs/images/log_parser_hld.png
