+++ internal/domain/log_mgmt/count_log_matches.go
